+++ assets/js/dashboard/stats/behaviours/conversions.js
@@ -9,14 +9,7 @@ import * as api from '../../api'
 import * as url from '../../util/url'
 import { escapeFilterValue } from '../../util/filters'
 import LazyLoader from '../../components/lazy-loader'
-
-function Money({ formatted }) {
-  if (formatted) {
-    return <span tooltip={formatted.long}>{formatted.short}</span>
-  } else {
-    return "-"
-  }
-}
+import Money from './money'
 
 export default class Conversions extends React.Component {
   constructor(props) {
@@ -40,7 +33,7 @@ export default class Conversions extends React.Component {
 
   componentDidUpdate(prevProps) {
     if (this.props.query !== prevProps.query) {
-      const height = this.htmlNode.current.element.offsetHeight
+      const height = this.htmlNode.current.offsetHeight
       this.setState({ loading: true, goals: null, prevHeight: height })
       this.fetchConversions()
     }
@@ -75,7 +68,7 @@ export default class Conversions extends React.Component {
             {renderRevenueColumn && <span className="hidden md:inline-block md:w-20 font-medium text-right"><Money formatted={goal.average_revenue} /></span>}
           </div>
         </div>
-        { renderProps && !goal.total_revenue && <PropBreakdown site={this.props.site} query={this.props.query} goal={goal} /> }
+        { renderProps && <PropBreakdown site={this.props.site} query={this.props.query} goal={goal} renderRevenueColumn={renderRevenueColumn } /> }
       </div>
     )
   }
@@ -108,9 +101,11 @@ export default class Conversions extends React.Component {
 
   renderConversions() {
     return (
-      <LazyLoader ref={this.htmlNode} style={{ minHeight: '132px', height: this.state.prevHeight ?? 'auto' }} onVisible={this.onVisible}>
-        {this.renderInner()}
-      </LazyLoader>
+      <div ref={this.htmlNode} style={{ minHeight: '132px', height: this.state.prevHeight ?? 'auto' }} >
+        <LazyLoader onVisible={this.onVisible}>
+          {this.renderInner()}
+        </LazyLoader>
+      </div>
     )
   }