+++ assets/js/dashboard/stats/locations/map.js
@@ -121,7 +121,7 @@ class Countries extends React.Component {
               this.props.query,
               {
                 country: country.code,
-                country_name: country.name
+                country_labels: country.name
               }
             )
           }